Changed projetos.index to display new data

+++ database/migrations/2024_03_25_100406_add_columns_to_project_table.php
@@ -11,7 +11,7 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('projetos', function (Blueprint $table) {
-            $table->integer('orcamento')->after('supervisor_id');
+            $table->double('orcamento')->after('supervisor_id');
             $table->string('tempoPrevisto')->after('orcamento');
             $table->string('tempoInvestido')->after('tempoPrevisto');
         });
